+++ tools/pubmed_tool.py
@@ -1,0 +1,7 @@
+"""
+Army81 Tools - PubMed wrapper
+يعيد التصدير من science_tools
+"""
+from tools.science_tools import search_pubmed
+
+__all__ = ["search_pubmed"]